pyproject.toml
[project]
name = "clip-engine"
version = "0.0.1"
description = ""
authors = [
    {name = "Gilderlan0101",email = "lansilva007gg@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "yt-dlp (>=2026.2.21,<2027.0.0)",
    "ffmpeg-python (>=0.2.0,<0.3.0)",
    "tqdm (>=4.67.3,<5.0.0)"
]

[tool.isort]
profile = "black"

# 🎯 CONFIGURAÇÃO DO BLACK
[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# 🎯 CONFIGURAÇÃO DO RUFF (alternativa moderna)
[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["src"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "fastapi[standard] (>=0.133.0,<0.134.0)",
    "blue (>=0.9.1,<0.10.0)",
    "isort (>=8.0.0,<9.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "tortoise-orm (>=1.1.5,<2.0.0)",
    "faster-whisper (>=1.2.1,<2.0.0)",
    "openai (>=2.24.0,<3.0.0)",
    "black (>=24.0.0,<25.0.0)",  # Adicionado!
    "ruff (>=0.3.0,<0.4.0)",
]
